requirements.txt
snowflake-sqlalchemy
snowflake-connector-python
apache-airflow
apache-airflow-providers-snowflake